requirements.txt
ultralytics
streamlit
opencv-python-headless
pillow
numpy
onnxruntime